Rename replyer to handler
